fix(tui): widen luma() accumulator to u32 to fix grayscale theme overflow panic
In debug builds, opening the `/theme` picker and previewing the Grayscale
theme on a cell carrying a bright RGB color (e.g. truecolor white) panics:

    thread '...' panicked at crates/tui/src/palette.rs:1141:7:
    attempt to add with overflow

The path is `adapt_*_for_grayscale_palette` → `grayscale_*_from_luma` →
`luma()`. The accumulator was a `u16`, but the weighted sum
`r*299 + g*587 + b*114` reaches `255 * 1000 = 255_000` for white, well
past `u16::MAX = 65_535`. Release builds wrap silently (which is its
own correctness problem — the returned `u8` was meaningless on overflow),
so the bug only surfaced as a panic in debug.

Fix: hold the accumulator in `u32`. The final result is still cast back
to `u8` after dividing by 1000, so the public contract is unchanged for
all non-overflowing inputs (the previous behavior on those inputs is
bit-identical). For inputs that previously overflowed, the new code now
returns the correct luma value instead of a wrapped one.

No test added because the function is `fn` (private) and any RGB above
the previous overflow threshold trivially exercises it through the
existing grayscale code paths.

+++ crates/tui/src/palette.rs
@@ -1138,7 +1138,10 @@ fn grayscale_bg_from_luma(luma: u8) -> Color {
 }
 
 fn luma(r: u8, g: u8, b: u8) -> u8 {
-    (((u16::from(r) * 299) + (u16::from(g) * 587) + (u16::from(b) * 114)) / 1000) as u8
+    // Coefficients sum to 1000; an RGB triple of (255,255,255) yields
+    // 255 * 1000 = 255_000 which overflows u16 (max 65_535). Use u32 to
+    // hold the weighted sum, then divide back into the 0..=255 range.
+    (((u32::from(r) * 299) + (u32::from(g) * 587) + (u32::from(b) * 114)) / 1000) as u8
 }
 // === Color depth + brightness helpers (v0.6.6 UI redesign) ===
